ref(backend): Simplify main.go, centralized error logging.
